+++ builder/parser/parser.go
@@ -61,6 +61,7 @@ func init() {
 		command.Expose:     parseStringsWhitespaceDelimited,
 		command.Volume:     parseMaybeJSONToList,
 		command.Insert:     parseIgnore,
+		command.Nocache:    parseString,
 	}
 }